Separated configFiles components
